refactored trace pool management, also removed cache flush routine as it has little effect in SMP systems;
