+++ libsandbox/src/platform.h
@@ -453,13 +453,6 @@ bool trace_end(const proc_t * const pproc);
  */
 void * sandbox_tracer(void * const psbox);
 
-/**
- * @brief Evict the existing blocks from the data caches.
- * This function was taken from Chapter 9 of Computer Systems A Programmer's
- * Perspective by Randal E. Bryant and David R. O'Hallaron
- */
-void cache_flush(void);
-
 #ifdef __cplusplus
 } /* extern "C" */
 #endif
